backend/requirements.txt
# Python version: 3.9+
# Tested on: Python 3.9

fastapi~=0.104.0
pydantic~=2.5.0
uvicorn~=0.24.0
Pillow~=10.1.0
httpx~=0.25.0
openai~=1.6.0
tenacity~=8.2.0
python-dotenv~=1.0.0
aiosqlite~=0.19.0
slowapi~=0.1.9
PyJWT~=2.8.0
zhdate~=0.1
qrcode~=7.4
dashscope~=1.14.0
cryptography~=41.0.0

# Dev / Test
pytest>=7.0
pytest-asyncio>=0.21
